// tmp/src/app/shared/enums/role.ts
export enum Role {
  BusinessSecurityAnalyst = 'analistasegempresarial',
  OperationalAnalyst = 'analistaoperacional',
  ControlCenter = 'centrocontrole',
  Consult = 'consulta',
  Transport = 'transporte',
  Paebm = 'paebm',
  Facilities = 'facilities',
  OtherAreas = 'demaisareas',
  Admin = 'admin',
  Euc = 'euc',
}